ADD & REF : added controllers and refactored followingrouter code to use the controllers
